Added continuous mode and created a SQLite database with it.
The continuous mode features a never ending climb for players. They are
to answer as many questions correctly as they can before missing a
question to achieve a high score. The SQLite database holds the value of
the score and an id to match it in case it needs to be edited or
removed. The Database returns an ArrayList of the scores that is
displayed in a ListView