Added music and added images for buttons.
The music works fine now. It may change. The music was handmade by me so
I'm not sure if its even any good. But its plays correctly and if I
choose it I will just need to replace the file. I also tried to add
background images to the answer choice and play and animation with them
but the background has severe runtime issues which I will need to fix. I
am going to try to change the margins of the questionBoxes to see if I
can mimic a scrolling look for nicer animations.